+++ UI/Components/ProgressBar.cs
@@ -158,7 +158,7 @@ namespace OTK.UI.Components
         /// <summary>
         /// Draws the progress bar to the screen if it is visible.
         /// </summary>
-        public override void Draw(bool ShareClipping = false)
+        public override void Draw()
         {
             if (!IsVisible) return;
             bool depthTestEnabled = GL.IsEnabled(EnableCap.DepthTest);