+++ UI/Components/Slider.cs
@@ -215,18 +215,18 @@ namespace OTK.UI.Components
         /// Draws the slider track and thumb to the screen.
         /// The thumb color is slightly dimmed while being dragged.
         /// </summary>
-        public override void Draw(bool ShareClipping = false)
+        public override void Draw()
         {
             if (!IsVisible) return;
             bool depthTestEnabled = GL.IsEnabled(EnableCap.DepthTest);
             bool blendEnabled = GL.IsEnabled(EnableCap.Blend);
             GL.Disable(EnableCap.DepthTest);
             GL.Enable(EnableCap.Blend);
-            base.Draw(ShareClipping);
+            base.Draw();
             thumb.Colour = _thumbColour;
             if (_isActive)
                 thumb.Colour = _thumbColour * 0.5f;
-            thumb.Draw(ShareClipping);
+            thumb.Draw();
             if (depthTestEnabled) GL.Enable(EnableCap.DepthTest);
             else GL.Disable(EnableCap.DepthTest);
             if (blendEnabled) GL.Enable(EnableCap.Blend);